Add a Sheet for Viewing by Google Analytics Prefix

+++ src/main/kotlin/com/si/main/searchinsights/data/PageViewInfo.kt
@@ -3,5 +3,5 @@ package com.si.main.searchinsights.data
 data class PageViewInfo(
     val pagePath: String,
     val pageTitle: String,
-    val pageViews: Int
+    val pageViews: Double
 )
